pyproject.toml
[project]
name = "trustcheck-agent"
version = "0.1.0"
description = "Multi-threaded website analysis agent for TrustCheck"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "google-genai>=1.0.0",
  "python-dotenv>=1.0.0",
]

[tool.uvicorn]
factory = false
